refactor: rename types into models. use interface instead of types for models. use nullable
