+++ lib/committee/index.ts
@@ -1,4 +1,4 @@
-export type { Committee, Link } from './types'
+export type { Committee, Link } from './models'
 export { Role, LinkType, ROLE_EMAILS } from './constants'
 
 export { LATEST_YEAR, SPONSOR_CONTACTS, TEAM } from './service'
